+++ taktik/core/social_media/instagram/ui/selectors/flows/unfollow.py
@@ -21,7 +21,55 @@ class UnfollowSelectors:
     unfollow_confirm_resource_id: str = 'com.instagram.android:id/primary_button'
     following_tab_text_probe: str = 'following'
     following_button_text: str = 'Following'
+    follow_back_button_text: str = 'Follow back'
     unfollow_confirm_text: str = 'Unfollow'
+    unified_follow_list_tab_layout_resource_name: str = 'unified_follow_list_tab_layout'
+    follow_list_subtitle_resource_name: str = 'follow_list_subtitle'
+
+    def active_resource_id(self, app_id: str, resource_name: str) -> str:
+        return f'{app_id}:id/{resource_name}'
+
+    def unified_follow_list_tab_layout_selector(self, app_id: str) -> str:
+        resource_id = self.active_resource_id(
+            app_id,
+            self.unified_follow_list_tab_layout_resource_name,
+        )
+        return f'//*[@resource-id="{resource_id}"]'
+
+    def unified_followers_tab_selector(self, app_id: str) -> str:
+        return (
+            self.unified_follow_list_tab_layout_selector(app_id)
+            + '//*[contains(@text, "Followers")]'
+        )
+
+    def active_follow_list_button_resource_id(self, app_id: str) -> str:
+        resource_name = self.following_list_button_resource_id.rsplit(':id/', 1)[-1]
+        return self.active_resource_id(app_id, resource_name)
+
+    def active_follow_list_username_resource_id(self, app_id: str) -> str:
+        resource_name = self.following_list_username_resource_id.rsplit(':id/', 1)[-1]
+        return self.active_resource_id(app_id, resource_name)
+
+    def active_follow_list_subtitle_resource_id(self, app_id: str) -> str:
+        return self.active_resource_id(app_id, self.follow_list_subtitle_resource_name)
+
+    def non_followers_category_selectors(self, app_id: str) -> List[str]:
+        return [
+            '//*[contains(@content-desc, "don\'t follow back")]',
+            '//*[contains(@content-desc, "People you don")]',
+            (
+                f'//*[@resource-id="{self.active_resource_id(app_id, "container")}"]'
+                '[contains(@content-desc, "follow")]'
+            ),
+            (
+                f'//*[@resource-id="{self.active_resource_id(app_id, "title")}"]'
+                '[contains(@text, "don\'t follow back")]'
+            ),
+            (
+                f'//*[@resource-id="{self.active_resource_id(app_id, "title")}"]'
+                '[contains(@text, "follow back")]'
+            ),
+        ]
     
     # === Confirmation d'unfollow dans la popup ===
     unfollow_confirm: List[str] = field(default_factory=lambda: [